+++ release_notes.md
@@ -96,6 +96,7 @@ First release into griffon artifacts portal.
 1. Add Swing builder 'glazedXXX()' node for generating JTable with GlazedLists.
 1. Add 'simple-jpa-console' script command for executing simple-jpa methods interactively.
 1. findModelByDsl() or findByDsl(Model) will return an entity or null while findAllModelByDsl() or findAllByDsl(Model) will return list entities.
+1. findModelBy() is replaced by findModelByAnd() or findByAnd(Model), findAllModelByAnd() or findAllByAnd(Model), findModelByOr() or findByOr(Model), findAllModelByOr() or findAllByOr(Model).
 
 TODO:
 1. Support finders such as findXXXByField1AndField2() or find XXXByField1OrField2AndField3().